requirement.txt
aiomysql==0.2.0
fastapi==0.115.6
mypy==1.13.0
pip==24.3.1
python-dotenv==1.0.1
python-telegram-bot==21.8
uvicorn==0.34.0
websockets==14.1
